refactor: Refactored optim module
