+++ holocron/optim/adamp.py
@@ -3,9 +3,11 @@
 # This program is licensed under the Apache License 2.0.
 # See LICENSE or go to <https://www.apache.org/licenses/LICENSE-2.0> for full license details.
 
-from typing import Callable, Iterable, Optional, Tuple
+import math
+from typing import Callable, Iterable, List, Optional, Tuple
 
 import torch
+from torch import Tensor
 from torch.optim import Adam
 
 from . import functional as F
@@ -119,7 +121,7 @@ class AdamP(Adam):
                     state_steps.append(state["step"])
 
             beta1, beta2 = group["betas"]
-            F.adamp(
+            adamp(
                 params_with_grad,
                 grads,
                 exp_avgs,
@@ -136,3 +138,55 @@ class AdamP(Adam):
             )
 
         return loss
+
+
+def adamp(
+    params: List[Tensor],
+    grads: List[Tensor],
+    exp_avgs: List[Tensor],
+    exp_avg_sqs: List[Tensor],
+    max_exp_avg_sqs: List[Tensor],
+    state_steps: List[int],
+    amsgrad: bool,
+    beta1: float,
+    beta2: float,
+    lr: float,
+    weight_decay: float,
+    eps: float,
+    delta: float,
+) -> None:
+    r"""Functional API that performs AdamP algorithm computation.
+    See :class:`~holocron.optim.AdamP` for details.
+    """
+
+    for i, param in enumerate(params):
+
+        grad = grads[i]
+        exp_avg = exp_avgs[i]
+        exp_avg_sq = exp_avg_sqs[i]
+        step = state_steps[i]
+
+        bias_correction1 = 1 - beta1**step
+        bias_correction2 = 1 - beta2**step
+
+        if weight_decay != 0:
+            grad = grad.add(param, alpha=weight_decay)
+
+        # Decay the first and second moment running average coefficient
+        exp_avg.mul_(beta1).add_(grad, alpha=1 - beta1)
+        exp_avg_sq.mul_(beta2).addcmul_(grad, grad, value=1 - beta2)
+        if amsgrad:
+            # Maintains the maximum of all 2nd moment running avg. till now
+            torch.maximum(max_exp_avg_sqs[i], exp_avg_sq, out=max_exp_avg_sqs[i])
+            # Use the max. for normalizing running avg. of gradient
+            denom = (max_exp_avg_sqs[i].sqrt() / math.sqrt(bias_correction2)).add_(eps)
+        else:
+            denom = (exp_avg_sq.sqrt() / math.sqrt(bias_correction2)).add_(eps)
+
+        # Extra step
+        pt = exp_avg / bias_correction1 / denom
+        if F.cosine_similarity(param.data.view(1, -1), grad.view(1, -1)).max() < delta / math.sqrt(param.data.numel()):
+            normalized_param = param.data / param.data.norm().add_(eps)
+            pt -= (normalized_param * pt).sum() * normalized_param.data
+
+        param.add_(pt, alpha=-lr)
