--- holocron/optim/functional.py
@@ -1,230 +1,0 @@
-# Copyright (C) 2019-2022, François-Guillaume Fernandez.
-
-# This program is licensed under the Apache License 2.0.
-# See LICENSE or go to <https://www.apache.org/licenses/LICENSE-2.0> for full license details.
-
-import math
-from typing import List
-
-import torch
-from torch import Tensor
-from torch.nn import functional as F
-
-
-def tadam(
-    params: List[Tensor],
-    grads: List[Tensor],
-    exp_avgs: List[Tensor],
-    exp_avg_sqs: List[Tensor],
-    max_exp_avg_sqs: List[Tensor],
-    W_ts: List[Tensor],
-    state_steps: List[int],
-    amsgrad: bool,
-    beta1: float,
-    beta2: float,
-    lr: float,
-    weight_decay: float,
-    eps: float,
-    dof: float,
-) -> None:
-    r"""Functional API that performs TAdam algorithm computation.
-    See :class:`~holocron.optim.TAdam` for details.
-    """
-
-    for i, param in enumerate(params):
-
-        grad = grads[i]
-        exp_avg = exp_avgs[i]
-        exp_avg_sq = exp_avg_sqs[i]
-        W_t = W_ts[i]
-        _dof = param.data.numel() if dof is None else dof
-        step = state_steps[i]
-        if amsgrad:
-            max_exp_avg_sq = max_exp_avg_sqs[i]
-
-        bias_correction1 = 1 - beta1**step
-        bias_correction2 = 1 - beta2**step
-
-        if weight_decay != 0:
-            grad = grad.add(param, alpha=weight_decay)
-
-        # Decay the first and second moment running average coefficient
-        w_t = grad.sub(exp_avg).pow_(2).div_(exp_avg_sq.add(eps)).sum()
-        w_t.add_(_dof).pow_(-1).mul_(_dof + param.data.numel())
-        exp_avg.mul_(W_t / (W_t + w_t)).addcdiv_(w_t * grad, W_t + w_t)
-        W_t.mul_((2 * beta1 - 1) / beta1)
-        W_t.add_(w_t)
-        exp_avg_sq.mul_(beta2).addcmul_(grad, grad, value=1 - beta2)
-        if amsgrad:
-            # Maintains the maximum of all 2nd moment running avg. till now
-            torch.maximum(max_exp_avg_sq, exp_avg_sq, out=max_exp_avg_sq)
-            # Use the max. for normalizing running avg. of gradient
-            denom = (max_exp_avg_sq.sqrt() / math.sqrt(bias_correction2)).add_(eps)
-        else:
-            denom = (exp_avg_sq.sqrt() / math.sqrt(bias_correction2)).add_(eps)
-
-        step_size = lr / bias_correction1
-
-        param.addcdiv_(exp_avg, denom, value=-step_size)
-
-
-def adabelief(
-    params: List[Tensor],
-    grads: List[Tensor],
-    exp_avgs: List[Tensor],
-    exp_avg_sqs: List[Tensor],
-    max_exp_avg_sqs: List[Tensor],
-    state_steps: List[int],
-    amsgrad: bool,
-    beta1: float,
-    beta2: float,
-    lr: float,
-    weight_decay: float,
-    eps: float,
-) -> None:
-    r"""Functional API that performs AdaBelief algorithm computation.
-    See :class:`~holocron.optim.AdaBelief` for details.
-    """
-
-    for i, param in enumerate(params):
-
-        grad = grads[i]
-        exp_avg = exp_avgs[i]
-        exp_avg_sq = exp_avg_sqs[i]
-        step = state_steps[i]
-        if amsgrad:
-            max_exp_avg_sq = max_exp_avg_sqs[i]
-
-        bias_correction1 = 1 - beta1**step
-        bias_correction2 = 1 - beta2**step
-
-        if weight_decay != 0:
-            grad = grad.add(param, alpha=weight_decay)
-
-        # Decay the first and second moment running average coefficient
-        exp_avg.mul_(beta1).add_(grad, alpha=1 - beta1)
-        grad_residual = grad - exp_avg
-        exp_avg_sq.mul_(beta2).addcmul_(grad_residual, grad_residual, value=1 - beta2)
-
-        if amsgrad:
-            # Maintains the maximum of all 2nd moment running avg. till now
-            torch.maximum(max_exp_avg_sq, exp_avg_sq, out=max_exp_avg_sq)
-            # Use the max. for normalizing running avg. of gradient
-            denom = (max_exp_avg_sq.sqrt() / math.sqrt(bias_correction2)).add_(eps)
-        else:
-            denom = (exp_avg_sq.sqrt() / math.sqrt(bias_correction2)).add_(eps)
-
-        step_size = lr / bias_correction1
-
-        param.addcdiv_(exp_avg, denom, value=-step_size)
-
-
-def adamp(
-    params: List[Tensor],
-    grads: List[Tensor],
-    exp_avgs: List[Tensor],
-    exp_avg_sqs: List[Tensor],
-    max_exp_avg_sqs: List[Tensor],
-    state_steps: List[int],
-    amsgrad: bool,
-    beta1: float,
-    beta2: float,
-    lr: float,
-    weight_decay: float,
-    eps: float,
-    delta: float,
-) -> None:
-    r"""Functional API that performs AdamP algorithm computation.
-    See :class:`~holocron.optim.AdamP` for details.
-    """
-
-    for i, param in enumerate(params):
-
-        grad = grads[i]
-        exp_avg = exp_avgs[i]
-        exp_avg_sq = exp_avg_sqs[i]
-        step = state_steps[i]
-
-        bias_correction1 = 1 - beta1**step
-        bias_correction2 = 1 - beta2**step
-
-        if weight_decay != 0:
-            grad = grad.add(param, alpha=weight_decay)
-
-        # Decay the first and second moment running average coefficient
-        exp_avg.mul_(beta1).add_(grad, alpha=1 - beta1)
-        exp_avg_sq.mul_(beta2).addcmul_(grad, grad, value=1 - beta2)
-        if amsgrad:
-            # Maintains the maximum of all 2nd moment running avg. till now
-            torch.maximum(max_exp_avg_sqs[i], exp_avg_sq, out=max_exp_avg_sqs[i])
-            # Use the max. for normalizing running avg. of gradient
-            denom = (max_exp_avg_sqs[i].sqrt() / math.sqrt(bias_correction2)).add_(eps)
-        else:
-            denom = (exp_avg_sq.sqrt() / math.sqrt(bias_correction2)).add_(eps)
-
-        # Extra step
-        pt = exp_avg / bias_correction1 / denom
-        if F.cosine_similarity(param.data.view(1, -1), grad.view(1, -1)).max() < delta / math.sqrt(param.data.numel()):
-            normalized_param = param.data / param.data.norm().add_(eps)
-            pt -= (normalized_param * pt).sum() * normalized_param.data
-
-        param.add_(pt, alpha=-lr)
-
-
-def adan(
-    params: List[Tensor],
-    grads: List[Tensor],
-    prev_grads: List[Tensor],
-    exp_avgs: List[Tensor],
-    exp_avg_sqs: List[Tensor],
-    exp_avg_deltas: List[Tensor],
-    max_exp_avg_deltas: List[Tensor],
-    state_steps: List[int],
-    amsgrad: bool,
-    beta1: float,
-    beta2: float,
-    beta3: float,
-    lr: float,
-    weight_decay: float,
-    eps: float,
-) -> None:
-    r"""Functional API that performs Adan algorithm computation.
-    See :class:`~holocron.optim.Adan` for details.
-    """
-
-    for i, param in enumerate(params):
-
-        grad = grads[i]
-        exp_avg = exp_avgs[i]
-        exp_avg_sq = exp_avg_sqs[i]
-        exp_avg_delta = exp_avg_deltas[i]
-        prev_grad = prev_grads[i]
-        step = state_steps[i]
-
-        bias_correction1 = 1 - beta1**step
-        bias_correction2 = 1 - beta2**step
-        bias_correction3 = 1 - beta3**step
-
-        if weight_decay != 0:
-            grad = grad.add(param, alpha=weight_decay)
-
-        # Decay the first and second moment running average coefficient
-        exp_avg.mul_(beta1).add_(grad, alpha=1 - beta1)
-        delta_grad = grad - prev_grad
-        exp_avg_sq.mul_(beta2).add_(delta_grad, alpha=1 - beta2)
-        _tmp = grad + beta2 * delta_grad
-        exp_avg_delta.mul_(beta3).addcmul_(_tmp, _tmp, value=1 - beta3)
-        if amsgrad:
-            # Maintains the maximum of all 2nd moment running avg. till now
-            torch.maximum(max_exp_avg_deltas[i], exp_avg_delta, out=max_exp_avg_deltas[i])
-            # Use the max. for normalizing running avg. of gradient
-            denom = (max_exp_avg_deltas[i].sqrt() / math.sqrt(bias_correction3)).add_(eps)
-        else:
-            denom = (exp_avg_delta.sqrt() / math.sqrt(bias_correction3)).add_(eps)
-
-        # Extra step
-        pt = (exp_avg / bias_correction1 + beta2 * exp_avg_sq / bias_correction2) / denom
-
-        param.add_(pt, alpha=-lr)
-        if weight_decay != 0:
-            param /= 1 + weight_decay * lr
