+++ holocron/optim/tadam.py
@@ -3,12 +3,12 @@
 # This program is licensed under the Apache License 2.0.
 # See LICENSE or go to <https://www.apache.org/licenses/LICENSE-2.0> for full license details.
 
-from typing import Callable, Dict, Iterable, Optional, Tuple
+import math
+from typing import Callable, Dict, Iterable, List, Optional, Tuple
 
 import torch
+from torch import Tensor
 from torch.optim.optimizer import Optimizer
-
-from . import functional as F
 
 
 class TAdam(Optimizer):
@@ -135,7 +135,7 @@ class TAdam(Optimizer):
                     # record the step after step update
                     state_steps.append(state["step"])
 
-            F.tadam(
+            tadam(
                 params_with_grad,
                 grads,
                 exp_avgs,
@@ -153,3 +153,60 @@ class TAdam(Optimizer):
             )
 
         return loss
+
+
+def tadam(
+    params: List[Tensor],
+    grads: List[Tensor],
+    exp_avgs: List[Tensor],
+    exp_avg_sqs: List[Tensor],
+    max_exp_avg_sqs: List[Tensor],
+    W_ts: List[Tensor],
+    state_steps: List[int],
+    amsgrad: bool,
+    beta1: float,
+    beta2: float,
+    lr: float,
+    weight_decay: float,
+    eps: float,
+    dof: float,
+) -> None:
+    r"""Functional API that performs TAdam algorithm computation.
+    See :class:`~holocron.optim.TAdam` for details.
+    """
+
+    for i, param in enumerate(params):
+
+        grad = grads[i]
+        exp_avg = exp_avgs[i]
+        exp_avg_sq = exp_avg_sqs[i]
+        W_t = W_ts[i]
+        _dof = param.data.numel() if dof is None else dof
+        step = state_steps[i]
+        if amsgrad:
+            max_exp_avg_sq = max_exp_avg_sqs[i]
+
+        bias_correction1 = 1 - beta1**step
+        bias_correction2 = 1 - beta2**step
+
+        if weight_decay != 0:
+            grad = grad.add(param, alpha=weight_decay)
+
+        # Decay the first and second moment running average coefficient
+        w_t = grad.sub(exp_avg).pow_(2).div_(exp_avg_sq.add(eps)).sum()
+        w_t.add_(_dof).pow_(-1).mul_(_dof + param.data.numel())
+        exp_avg.mul_(W_t / (W_t + w_t)).addcdiv_(w_t * grad, W_t + w_t)
+        W_t.mul_((2 * beta1 - 1) / beta1)
+        W_t.add_(w_t)
+        exp_avg_sq.mul_(beta2).addcmul_(grad, grad, value=1 - beta2)
+        if amsgrad:
+            # Maintains the maximum of all 2nd moment running avg. till now
+            torch.maximum(max_exp_avg_sq, exp_avg_sq, out=max_exp_avg_sq)
+            # Use the max. for normalizing running avg. of gradient
+            denom = (max_exp_avg_sq.sqrt() / math.sqrt(bias_correction2)).add_(eps)
+        else:
+            denom = (exp_avg_sq.sqrt() / math.sqrt(bias_correction2)).add_(eps)
+
+        step_size = lr / bias_correction1
+
+        param.addcdiv_(exp_avg, denom, value=-step_size)
